✨ Create functional jump checks
So, the layermask is not exactly behaving as I expected, and it does warrant some extra investigation. But, despite the less than clean implementation (you need to adjust the jumpable layers within the boxcollider2D), the player can no longer double jump!

I'll make an issue regarding falling off edges, as it can be a bit annoying. It's also 2AM here.