⚗️ Discover why the layermask was acting strange
Seems logical in hindsight, but I'd have imagined unity would have built some easy bitwise function for this. For now, you need to bitshift the layer by one to the left, as to make it correspond to the position in the bitmask. Might want to write a small bit of code for this at some point, but not now. Explains #13